+++ app/models/group.rb
@@ -28,6 +28,11 @@ class Group < ApplicationRecord # rubocop:disable ClassLength
   belongs_to :lottery_assignment
   accepts_nested_attributes_for :suite
 
+  # This before_destroy clears room assignments and needs to be added before
+  # the dependent: :delete_all in the memberships association or there won't be
+  # any members to update
+  before_destroy :remove_member_rooms
+
   has_many :memberships, dependent: :delete_all
   has_many :full_memberships, -> { where(status: 'accepted') },
            class_name: 'Membership', inverse_of: :group
@@ -61,7 +66,6 @@ class Group < ApplicationRecord # rubocop:disable ClassLength
                 if: -> { will_save_change_to_lottery_assignment_id? }
   after_update :remove_clip_memberships,
                if: ->() { changed_draw_with_clip_memberships? }
-  before_destroy :remove_member_rooms
   after_destroy :restore_member_draws, if: ->(g) { g.draw.nil? }
   after_destroy :notify_members_of_disband
   after_destroy :destroy_lottery_assignment,